winc1500: nmspi: Don't force definition of USE_OLD_SPI_SW
If this macro is forced here SPI driver code uses separate Write and
read back for bulk writes and fails doing so. If this is not defined SPI
read/write cycle is used and it perfomes successfully. This could be due
to a undocumented change in the WiFi module firmware in the newer
modules. Did not check with older module (firmware) since I did not have
them around.
Instead, we make this configurable via Kconfig on Zephyr side
and pass as a build option to this module. So that the option is
available without having to dig into the driver code.
This driver was tested along with nRF52840 and nRF52382 DK
boards and found that this define, if forced, will not work.
Two boards not working is enough to get this define moved
to Kconfig.
The corresponding changes in Zephyr are in
zephyrproject-rtos/zephyr#26907

Signed-off-by: Raja D.Singh <rdsingh@iotwizards.com>