pinctrl: Add wake-up sources generation
The SAM SoC requires that SUPC be configured to enable specific pins
to be used as wake-up sources. This add a new section denominated as
wakeup. In that section the pinmux will be used to store the correct
wkupX input to be used by the pinctrl driver to configure the gpio
and supc accordingly.

This update all pinconfigs, the pinctrl definitions, update the
documentation  and expand the current test to cover the changes.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>