Macro rename for samv71
The pin macro for SAM V71 Q relative to PE5 is named as if it is
relative to AFEC1 AD3 (PE5X_AFE1_AD3). Same typo can be found for the
pin PE4 for the AFE1 AD4 that is AFE0 AD4.

The pinconfig YAML for the s70, e70 and v7x has a typo for the AFEC.
The typo is this: AFEC0 AD3 and AD4 was referenced for AFEC1 instead of
AFE0.  The solution proposed in this commit is:
- Edit the pinconfigs YAML for [s,e]70, v7x.
- Run the sampinconfig.py script to regenerate the includes

Signed-off-by: zebo9x <ivan.zebo@gmail.com>